+++ rpc/src/main/java/com/walmartlabs/concord/rpc/Command.java
@@ -1,4 +1,6 @@
 package com.walmartlabs.concord.rpc;
 
 public interface Command {
+
+    CommandType getType();
 }