+++ rpc/src/main/java/com/walmartlabs/concord/rpc/CommandType.java
@@ -1,0 +1,6 @@
+package com.walmartlabs.concord.rpc;
+
+public enum CommandType {
+
+    CANCEL_JOB
+}