+++ server/impl/src/main/java/com/walmartlabs/concord/server/template/TemplateProcessor.java
@@ -2,7 +2,7 @@ package com.walmartlabs.concord.server.template;
 
 import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.project.Constants;
-import com.walmartlabs.concord.server.LogManager;
+import com.walmartlabs.concord.server.process.logs.LogManager;
 import com.walmartlabs.concord.server.metrics.WithTimer;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;